Add test for PEP 495 gap logic in tzfile
PEP 495 says that when faced with an ambiguous choice (in a fold or a
gap), a tzinfo should return the offset from *before* the transition
when fold is 0 and the offset from *after* the transition when fold is
1, which affects both gaps (imaginary times) and folds (ambiguous
times).

Currently, `dateutil.tz.tzfile` only properly handles ambiguous times,
not imaginary times.

This adds xfailing tests for the proper handling of transitions with a
gap.